+++ test/test_connectionpool.py
@@ -92,13 +92,21 @@ class TestConnectionPool(unittest.TestCase):
 
     def test_retry_exception_str(self):
         self.assertEqual(
-            str(MaxRetryError(HTTPConnectionPool(host='localhost'), "Test.", None)),
-            "HTTPConnectionPool(host='localhost', port=None): Max retries exceeded with url: Test. (Caused by redirect)")
+            str(MaxRetryError(
+                HTTPConnectionPool(host='localhost'), "Test.", None)),
+            "HTTPConnectionPool(host='localhost', port=None): "
+            "Max retries exceeded with url: Test. (Caused by redirect)")
 
         err = SocketError("Test")
+
+        # using err.__class__ here, as socket.error is an alias for OSError
+        # since Py3.3 and gets printed as this
         self.assertEqual(
-            str(MaxRetryError(HTTPConnectionPool(host='localhost'), "Test.", err)),
-            "HTTPConnectionPool(host='localhost', port=None): Max retries exceeded with url: Test. (Caused by <class 'socket.error'>: Test)")
+            str(MaxRetryError(
+                HTTPConnectionPool(host='localhost'), "Test.", err)),
+            "HTTPConnectionPool(host='localhost', port=None): "
+            "Max retries exceeded with url: Test. "
+            "(Caused by {0}: Test)".format(str(err.__class__)))
 
     def test_pool_size(self):
         POOL_SIZE = 1
@@ -109,8 +117,7 @@ class TestConnectionPool(unittest.TestCase):
 
         def _test(exception, expect):
             pool._make_request = lambda *args, **kwargs: _raise(exception)
-            with self.assertRaises(expect):
-                pool.request('GET', '/')
+            self.assertRaises(expect, pool.request, 'GET', '/')
 
             self.assertEqual(pool.pool.qsize(), POOL_SIZE)
 
@@ -125,15 +132,15 @@ class TestConnectionPool(unittest.TestCase):
         # MaxRetryError, not EmptyPoolError
         # See: https://github.com/shazow/urllib3/issues/76
         pool._make_request = lambda *args, **kwargs: _raise(HTTPException)
-        with self.assertRaises(MaxRetryError):
-            pool.request('GET', '/', retries=1, pool_timeout=0.01)
+        self.assertRaises(MaxRetryError, pool.request,
+                          'GET', '/', retries=1, pool_timeout=0.01)
         self.assertEqual(pool.pool.qsize(), POOL_SIZE)
 
     def test_assert_same_host(self):
         c = connection_from_url('http://google.com:80')
 
-        with self.assertRaises(HostChangedError):
-            c.request('GET', 'http://yahoo.com:80', assert_same_host=True)
+        self.assertRaises(HostChangedError, c.request,
+                          'GET', 'http://yahoo.com:80', assert_same_host=True)
 
     def test_pool_close(self):
         pool = connection_from_url('http://google.com:80')
@@ -150,16 +157,13 @@ class TestConnectionPool(unittest.TestCase):
         pool.close()
         self.assertEqual(pool.pool, None)
 
-        with self.assertRaises(ClosedPoolError):
-            pool._get_conn()
+        self.assertRaises(ClosedPoolError, pool._get_conn)
 
         pool._put_conn(conn3)
 
-        with self.assertRaises(ClosedPoolError):
-            pool._get_conn()
+        self.assertRaises(ClosedPoolError, pool._get_conn)
 
-        with self.assertRaises(Empty):
-            old_pool_queue.get(block=False)
+        self.assertRaises(Empty, old_pool_queue.get, block=False)