+++ test/test_filepost.py
@@ -101,6 +101,20 @@ class TestMultipartEncoding(unittest.TestCase):
             b'multipart/form-data; boundary=' + b(BOUNDARY))
 
 
+    def test_explicit(self):
+        fields = [('k', ('somefile.txt', b'v', 'image/jpeg'))]
+
+        encoded, content_type = encode_multipart_formdata(fields, boundary=BOUNDARY)
+
+        self.assertEqual(encoded,
+            b'--' + b(BOUNDARY) + b'\r\n'
+            b'Content-Disposition: form-data; name="k"; filename="somefile.txt"\r\n'
+            b'Content-Type: image/jpeg\r\n'
+            b'\r\n'
+            b'v\r\n'
+            b'--' + b(BOUNDARY) + b'--\r\n'
+            )
+
     def test_international_headers(self):
         fields = [(u'ke\u00ff', (u'n\u00e4me', b'v1')),
                   (u'\u03a4\u03b5\u03be\u03c4', b'v2')]