+++ test/test_response.py
@@ -31,6 +31,12 @@ class TestResponse(unittest.TestCase):
         r = HTTPResponse(None)
         self.assertEqual(r.data, None)
 
+    def test_release_conn(self):
+        r = HTTPResponse(None)
+        r._fp = BytesIO()
+        r.release_conn()
+        self.assertTrue(r._fp.closed)
+
     def test_preload(self):
         fp = BytesIO(b'foo')