+++ test/with_dummyserver/test_socketlevel.py
@@ -1,10 +1,15 @@
-from urllib3 import HTTPConnectionPool
+from urllib3 import HTTPConnectionPool, HTTPSConnectionPool
 from urllib3.poolmanager import proxy_from_url
-from urllib3.exceptions import MaxRetryError, TimeoutError
+from urllib3.exceptions import MaxRetryError, TimeoutError, SSLError
 
 from dummyserver.testcase import SocketDummyServerTestCase
 
 from threading import Event
+
+try:
+    from ssl import HAS_SNI
+except ImportError: # openssl without SNI
+    HAS_SNI = False
 
 
 class TestCookies(SocketDummyServerTestCase):
@@ -26,6 +31,30 @@ class TestCookies(SocketDummyServerTestCase):
         pool = HTTPConnectionPool(self.host, self.port)
         r = pool.request('GET', '/', retries=0)
         self.assertEquals(r.headers, {'set-cookie': 'foo=1, bar=1'})
+
+
+if HAS_SNI:
+    class TestSNI(SocketDummyServerTestCase):
+
+        def test_hostname_in_first_request_packet(self):
+            done_receiving = Event()
+            self.buf = b''
+
+            def socket_handler(listener):
+                sock = listener.accept()[0]
+
+                self.buf = sock.recv(65536) # We only accept one packet
+                done_receiving.set()  # let the test know it can proceed
+
+            self._start_server(socket_handler)
+            pool = HTTPSConnectionPool(self.host, self.port)
+            try:
+                pool.request('GET', '/', retries=0)
+            except SSLError: # We are violating the protocol
+                pass
+            done_receiving.wait()
+            self.assertTrue(self.host.encode() in self.buf,
+                            "missing hostname in SSL handshake")
 
 
 class TestSocketClosing(SocketDummyServerTestCase):
@@ -72,8 +101,7 @@ class TestSocketClosing(SocketDummyServerTestCase):
         # Does the pool retry if there is no listener on the port?
         # Note: Socket server is not started until after the test.
         pool = HTTPConnectionPool(self.host, self.port)
-        with self.assertRaises(MaxRetryError):
-            pool.request('GET', '/')
+        self.assertRaises(MaxRetryError, pool.request, 'GET', '/')
         self._start_server(lambda x: None)
 
     def test_connection_timeout(self):
@@ -86,8 +114,7 @@ class TestSocketClosing(SocketDummyServerTestCase):
         self._start_server(socket_handler)
         pool = HTTPConnectionPool(self.host, self.port, timeout=0.001)
 
-        with self.assertRaises(TimeoutError):
-            pool.request('GET', '/', retries=0)
+        self.assertRaises(TimeoutError, pool.request, 'GET', '/', retries=0)
 
         timed_out.set()